Fix sign promotion warnings in NSharedLinux::GetClockTicks.
As the clock is always positive we can safely promote to uint64_t.

+++ Library/Source/Nano/Internal/Targets/Shared/NSharedLinux.cpp
@@ -89,7 +89,7 @@ uint64_t NSharedLinux::GetClockTicks()
 		memset(&timeSpec, 0x00, sizeof(timeSpec));
 	}
 
-	return (timeSpec.tv_sec * kNanosecondsPerSecond) + timeSpec.tv_nsec;
+	return (uint64_t(timeSpec.tv_sec) * kNanosecondsPerSecond) + uint64_t(timeSpec.tv_nsec);
 }
